Fix size_t vs int

+++ yt/server/query_agent/query_executor.cpp
@@ -221,7 +221,7 @@ private:
             writer,
             false,
             ranges,
-            [&] (const TConstQueryPtr& subquery, size_t index) {
+            [&] (const TConstQueryPtr& subquery, int index) {
                 std::vector<ISchemafulReaderPtr> bottomSplitReaders;
                 for (const auto& dataSplit : groupedSplits[index]) {
                     bottomSplitReaders.push_back(GetReader(dataSplit, nodeDirectory));
